Remove nodejs requirement for host machine
Silent failures generated a lot of issues, so I moved the node and npm
stuff to the web containers. The files are now just copied to the web container
instead of being mounted as a volume, which means that some changes are needed
for active development (which requires nodejs on the host machine).
Will update instructions if people ask for it.